+++ xla/backends/gpu/runtime/command_state_test.cc
@@ -1,4 +1,4 @@
-/* Copyright 2025 The OpenXLA Authors.
+/* Copyright 2026 The OpenXLA Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -13,9 +13,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
+#include "xla/backends/gpu/runtime/command_state.h"
+
 #include <cstdint>
 
-#include "xla/backends/gpu/runtime/command_state.h"
 #include "xla/tsl/platform/test.h"
 
 namespace xla::gpu {
@@ -31,7 +32,7 @@ struct StateB : public CommandState {
 
 TEST(CommandStateManagerTest, GetOrCreateState) {
   // We need a fake command pointer to use as a key. Nullptr works just fine!
-  const CommandBufferCmd* cmd = nullptr;
+  const Command* cmd = nullptr;
 
   CommandStateManager state_manager(/*command_buffer=*/nullptr);
